change motor gear ratio int -> float

+++ robomas_encoder/robomas_encoder.hpp
@@ -19,7 +19,7 @@ class robomas_encoder{
 private:
     uint8_t robomas_input_rpm_high = 0, robomas_input_rpm_low = 0; //save robomas_encoder_data
     uint8_t robomas_input_pos_high = 0, robomas_input_pos_low = 0;
-    int16_t debug_robomas_rpm = 0;
+    float debug_robomas_rpm = 0;
     int16_t signed_robomas_pos_data = 0;
     float signed_robomas_pos_axis_data = 0.0f;
     float debug_robomas_pos = 0.0f;
@@ -36,7 +36,7 @@ public:
 
     void change_motor_type(robomas_motor_type_enum motor_type);
     void input_encoder_data(uint8_t[8]);
-    int show_rpm();
+    float show_rpm();
     float show_pos();
     void reset_positon(int new_default_position = 0);
